implementation of bedevere package
Removed the markov and discrete_sums modules and implemented them with the pypi "bedevere" module. The code contained within them was OSRS independent and I will henceforth maintain them as I need, both for this project and others.

A link to the bedevere package is given: https://pypi.org/project/bedevere/
The code can be found on github as well: https://github.com/noahgill409/bedevere